# server/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "description": "Body of every non-2xx response.",
        "properties": {
          "error": {
            "type": "string"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "TokenResponse": {
        "properties": {
          "token": {
            "type": "string"
          }
        },
        "required": [
          "token"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "description": "The server's configured `api_key`.",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "The Muorg music server HTTP API, shared by the desktop (Tauri), web and Android clients. Everything except `GET /api/health` and `GET /stream/{id}` requires an `Authorization: Bearer <api_key>` header.\n\n`GET /api/tracks` is paginated and defaults to 500 rows — a bare call returns the first page, not the catalog. Follow `X-Total-Count`.",
    "license": {
      "name": ""
    },
    "title": "Muorg Server API",
    "version": "2.42.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/tracks/{id}/stream-token": {
      "get": {
        "operationId": "issue_token",
        "parameters": [
          {
            "description": "Track id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TokenResponse"
                }
              }
            },
            "description": "Stream token"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Missing or invalid API key"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "No such track"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "Mint a short-lived (8h) token for the public `/stream/{id}` URL, so an\n`<audio src>` can play without carrying the API key.",
        "tags": [
          "Stream"
        ]
      }
    },
    "/stream/{id}": {
      "get": {
        "operationId": "stream_audio",
        "parameters": [
          {
            "description": "Track id",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          },
          {
            "description": "Short-lived token from `/api/tracks/{id}/stream-token`.",
            "in": "query",
            "name": "token",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "description": "Seek offset in seconds; the server transcodes from this point.",
            "in": "query",
            "name": "start",
            "required": false,
            "schema": {
              "format": "float",
              "type": [
                "number",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "audio/mpeg": {}
            },
            "description": "Audio bytes"
          },
          "206": {
            "description": "Partial content for a ranged request"
          },
          "401": {
            "description": "Missing, expired or wrong token"
          },
          "404": {
            "description": "No such track"
          }
        },
        "summary": "Audio bytes, range-capable. Public: authorised by the `token` query\nparameter rather than the Bearer header, so browsers and cast receivers can\nfetch it directly.",
        "tags": [
          "Stream"
        ]
      }
    }
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
}